+++ artifacts/predefined.yaml
@@ -8,3 +8,4 @@ bad_accounts:
   - drivemodesamsungnote20@gmail.com
 
 default_recipient: "laurence@drivemode.com"
+collection_id: 6f6295c9-d42a-407b-a721-ad27008e96b9
